requirements.txt
aiohttp==3.9.3 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
appnope==0.1.3 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Darwin"
asttokens==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.10" and python_version < "3.11"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.10" and python_version < "4.0"
bs4==0.0.2 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.3.2 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0" and implementation_name == "pypy"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
comm==0.2.1 ; python_version >= "3.10" and python_version < "4.0"
contourpy==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
db-dtypes==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
debugpy==1.8.0 ; python_version >= "3.10" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
executing==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
fastjsonschema==2.19.1 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.47.2 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2024.2.0 ; python_version >= "3.10" and python_version < "4.0"
gcsfs==2024.2.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.17.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.17.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-oauthlib==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.27.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-bigquery-storage==2.24.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-bigquery==3.17.2 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-secret-manager==2.18.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.14.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud==0.34.0 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.62.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos[grpc]==1.62.0 ; python_version >= "3.10" and python_version < "4.0"
grpc-google-iam-v1==0.13.0 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.60.1 ; python_version < "4.0" and python_version >= "3.10"
grpcio==1.60.1 ; python_version < "4.0" and python_version >= "3.10"
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
ipykernel==6.28.0 ; python_version >= "3.10" and python_version < "4.0"
ipython==8.21.0 ; python_version >= "3.10" and python_version < "4.0"
jedi==0.19.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.10" and python_version < "4.0"
jupyter-client==8.6.0 ; python_version >= "3.10" and python_version < "4.0"
jupyter-core==5.7.1 ; python_version >= "3.10" and python_version < "4.0"
kaleido==0.2.1 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.10" and python_version < "4.0"
matplotlib-inline==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.8.2 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.10" and python_version < "4.0"
nbformat==5.9.2 ; python_version >= "3.10" and python_version < "4.0"
nest-asyncio==1.6.0 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.26.3 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.10" and python_version < "4.0"
pandas-gbq==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.1.4 ; python_version >= "3.10" and python_version < "4.0"
parso==0.8.3 ; python_version >= "3.10" and python_version < "4.0"
pexpect==4.9.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "win32"
pillow==10.2.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==4.2.0 ; python_version >= "3.10" and python_version < "4.0"
plotly==5.18.0 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.43 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.25.2 ; python_version >= "3.10" and python_version < "4.0"
psutil==5.9.8 ; python_version >= "3.10" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "win32"
pure-eval==0.2.2 ; python_version >= "3.10" and python_version < "4.0"
pyarrow==15.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0" and implementation_name == "pypy"
pydata-google-auth==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.17.2 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.4 ; python_version >= "3.10" and python_version < "4.0"
pywin32==306 ; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
pyzmq==25.1.2 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.33.0 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
retrying==1.3.4 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.17.1 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
setuptools==69.1.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.10" and python_version < "4.0"
stack-data==0.6.3 ; python_version >= "3.10" and python_version < "4.0"
tenacity==8.2.3 ; python_version >= "3.10" and python_version < "4.0"
tornado==6.4 ; python_version >= "3.10" and python_version < "4.0"
traitlets==5.14.1 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.4 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.10" and python_version < "4.0"
